requirements.txt
amqp==5.1.1
gunicorn
redis
psycopg2-binary==2.9.1
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.4
django-ckeditor-5==0.2.2
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-jalali==6.0.1
django-jalali-date==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
isodate==0.6.1
jdatetime==4.1.0
kavenegar==1.1.2
kombu==5.2.4
lxml==4.9.2
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
pycodestyle==2.10.0
pycryptodome==3.16.0
PyJWT==2.6.0
pytz==2022.7
requests==2.28.1
requests-file==1.5.1
requests-toolbelt==0.10.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
zeep==4.2.1
